move build to be a method of the model
