+++ kornia/contrib/__init__.py
@@ -1,3 +1,5 @@
+from kornia.contrib.sam import Sam, SamModelType
+
 from .classification import ClassificationHead
 from .connected_components import connected_components
 from .diamond_square import diamond_square
@@ -35,4 +37,6 @@ __all__ = [
     "distance_transform",
     "DistanceTransform",
     "diamond_square",
+    "Sam",
+    "SamModelType",
 ]
