+++ kornia/contrib/sam/__init__.py
@@ -1,3 +1,3 @@
-from kornia.contrib.sam.model import SamType
-from kornia.contrib.sam.model import load_sam as load
-from kornia.contrib.sam.predictor import SamPredictor
+from kornia.contrib.sam.architecture import Sam, SamModelType
+
+# from kornia.contrib.sam.predictor import SamPredictor
